+++ squirrel/src/main/java/ua/squirrel/user/entity/employee/EmployeeModel.java
@@ -13,7 +13,7 @@ public class EmployeeModel {
 
 	private String lastName;
 
-	private Long salary;
+	private Integer salary;
 
 	private Long storeId;
